Move ParserOptionsEditor under data table
