+++ javascript/app-framework/src/components/editors/DataTableEditor/DataTableEditor.tsx
@@ -5,8 +5,8 @@
 import { memo } from 'react'
 
 import { JsonModeEditor } from '../JsonModeEditor/index.js'
-import { ParserOptionsEditor } from '../ParserOptionsEditor/index.js'
 import type { DataTableEditorProps } from './DataTableEditor.types.js'
+import { ParserOptionsEditor } from './ParserOptionsEditor/index.js'
 
 export const DataTableEditor: React.FC<DataTableEditorProps> = memo(
 	function DataTableEditor({ resource }) {
