+++ javascript/app-framework/src/components/editors/DataTableEditor/ParserOptionsEditor/ParserOptionsEditor.hooks.ts
