+++ javascript/app-framework/src/components/editors/DataTableEditor/ParserOptionsEditor/ParserOptionsEditor.styles.ts
