+++ javascript/app-framework/src/components/editors/DataTableEditor/ParserOptionsEditor/index.ts
