+++ javascript/app-framework/src/components/editors/index.ts
@@ -4,8 +4,8 @@
  */
 export * from './CodebookEditor/index.js'
 export * from './DataTableEditor/index.js'
+export * from './DataTableEditor/ParserOptionsEditor/index.js'
 export * from './JsonEditor/index.js'
-export * from './ParserOptionsEditor/index.js'
 export * from './RawTableViewer/index.js'
 export * from './TableBundleEditor/index.js'
 export * from './ViewOptions/index.js'
